ref(slack): migrate slack formatter out of platforms
