--- app/infrastructure/platforms/formatters/__init__.py
@@ -1,13 +1,0 @@
-"""Response formatters for platform-specific message formats (Block Kit, Adaptive Cards, Embeds)."""
-
-from infrastructure.platforms.formatters.base import BaseResponseFormatter
-from infrastructure.platforms.formatters.discord import DiscordEmbedFormatter
-from infrastructure.platforms.formatters.slack import SlackBlockKitFormatter
-from infrastructure.platforms.formatters.teams import TeamsAdaptiveCardsFormatter
-
-__all__ = [
-    "BaseResponseFormatter",
-    "SlackBlockKitFormatter",
-    "TeamsAdaptiveCardsFormatter",
-    "DiscordEmbedFormatter",
-]
