+++ app/infrastructure/platforms/providers/slack.py
@@ -18,10 +18,10 @@ from infrastructure.platforms.capabilities.models import (
     PlatformCapability,
     create_capability_declaration,
 )
-from infrastructure.platforms.formatters.slack import SlackBlockKitFormatter
 from infrastructure.platforms.parsing import CommandArgumentParser
 from infrastructure.platforms.providers.base import BasePlatformProvider
 from integrations.slack import LegacySlackBootstrap
+from integrations.slack.formatter import SlackBlockKitFormatter
 from integrations.slack.help import (
     SLACK_HELP_KEYWORDS,
     SlackHelpGenerator,
