+++ app/tests/unit/infrastructure/platforms/test_slack_auto_registration.py
@@ -13,8 +13,8 @@ import pytest
 from infrastructure.configuration.infrastructure.platforms import (
     SlackPlatformSettings,
 )
-from infrastructure.platforms.formatters.slack import SlackBlockKitFormatter
 from infrastructure.platforms.providers.slack import SlackPlatformProvider
+from integrations.slack.formatter import SlackBlockKitFormatter
 from integrations.slack.models import CommandPayload, CommandResponse
